refactor: move test to pipeline crate
