+++ Cargo.toml
@@ -20,26 +20,31 @@ ansi_term = "0.12"
 anyhow = "1"
 axum = { version = "0.6", features = ["ws"] }
 bytes = "1"
-cargo-lock = "9"
-cargo_metadata = "0.15"
 clap = { version = "4", features = ["derive", "env"] }
 console = "0.15"
 dunce = "1"
 envy = "0.4"
 fs_extra = "1"
-futures-util = { version = "0.3", default-features = false, features = ["sink"] }
+futures-util = { version = "0.3", default-features = false, features = [
+    "sink",
+] }
 local-ip-address = "0.5.1"
 nipper = "0.1"
 notify = "4"
 once_cell = "1"
 open = "5"
-remove_dir_all = "0.8"
-reqwest = { version = "0.11", default-features = false, features = ["rustls-tls", "stream", "trust-dns"] }
-seahash = "4"
+reqwest = { version = "0.11", default-features = false, features = [
+    "rustls-tls",
+    "stream",
+    "trust-dns",
+] }
 serde = { version = "1", features = ["derive"] }
 # See https://docs.rs/tokio/latest/tokio/#feature-flags - we basically use all of the features.
 tokio = { version = "1", default-features = false, features = ["full"] }
-tokio-stream = { version = "0.1", default-features = false, features = ["fs", "sync"] }
+tokio-stream = { version = "0.1", default-features = false, features = [
+    "fs",
+    "sync",
+] }
 tokio-tungstenite = "0.19"
 toml = "0.7"
 tower-http = { version = "0.4", features = ["fs", "trace"] }
